homework #12 Spring 2
unfinished project.
minor fixes.
1. finished changing RestTemplate on WebClient
2. Replaced double type at price and total price

+++ common/src/main/java/ru/gb/sklyarov/shop/common/dtos/CartDto.java
@@ -1,15 +1,16 @@
 package ru.gb.sklyarov.shop.common.dtos;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public class CartDto {
     private List<OrderItemDto> cartItems;
-    private double totalPrice;
+    private BigDecimal totalPrice;
 
     public CartDto() {
     }
 
-    public CartDto(List<OrderItemDto> cartItems, double totalPrice) {
+    public CartDto(List<OrderItemDto> cartItems, BigDecimal totalPrice) {
         this.cartItems = cartItems;
         this.totalPrice = totalPrice;
     }
@@ -22,11 +23,11 @@ public class CartDto {
         this.cartItems = cartItems;
     }
 
-    public double getTotalPrice() {
+    public BigDecimal getTotalPrice() {
         return totalPrice;
     }
 
-    public void setTotalPrice(double totalPrice) {
+    public void setTotalPrice(BigDecimal totalPrice) {
         this.totalPrice = totalPrice;
     }
 }
