+++ core-service/src/main/java/ru/gb/sklyarov/shop/core/ms/repositories/ProductRepository.java
@@ -4,15 +4,16 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import ru.gb.sklyarov.shop.core.ms.entities.Product;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Repository
 public interface ProductRepository extends JpaRepository<Product, Long> {
-    List<Product> findAllByPriceIsLessThanEqual(Double maxPriceLimit);
+    List<Product> findAllByPriceIsLessThanEqual(BigDecimal maxPriceLimit);
 
-    List<Product> findAllByPriceIsGreaterThanEqual(Double minPriceLimit);
+    List<Product> findAllByPriceIsGreaterThanEqual(BigDecimal minPriceLimit);
 
-    List<Product> findAllByPriceBetween(Double minPriceLimit, Double maxPriceLimit);
+    List<Product> findAllByPriceBetween(BigDecimal minPriceLimit, BigDecimal maxPriceLimit);
 
     // TODO: переделать на rest
 //    @Query(value = "select count(distinct ord.id) from orders ord inner join order_items oi on oi.order_id = ord.id" +
